Created a new action which will be used to serve all requests passed to the builder service.
The new action serves both compilation and library-info-fetching requests. Depending on the request type ('compiler' or 'library'),
the main action will call the respective function and the pass the response to the caller.
Every request should have the correct API settings (auth-key and API version) and must include the must include the type of
request in the body. Every other combination of these will be rejected.