Added 'header-check' option to the builder API
- Implemented new API method
- Added functional tests

The new API method is primarily meant to be used during library testing.
The method is capable of detecting whether a specified header is
actually used in a project's code. If the specified header is 'included'
in the projects .ino file, but exists in the project files, it is
considered by the method as not being used.